requirements.txt
absl-py==0.9.0
astunparse==1.6.3
audioread==2.1.8
cachetools==4.1.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
config==0.4.2
cycler==0.10.0
decorator==4.4.2
gast==0.3.3
google-auth==1.15.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.29.0
h5py==2.10.0
idna==2.9
imagecodecs==2020.2.18
imageio==2.8.0
joblib==0.15.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.1.0
librosa==0.7.2
llvmlite==0.31.0
Markdown==3.2.2
matplotlib==3.2.0
mutagen==1.44.0
networkx==2.4
numba==0.48.0
numpy==1.18.1
oauthlib==3.1.0
opt-einsum==3.2.1
pandas==1.0.3
Pillow==7.1.2
Pillow-PIL==0.1.dev0
protobuf==3.12.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydub==0.23.1
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.0
scikit-image==0.17.2
scikit-learn==0.23.0
scipy==1.4.1
six==1.14.0
SoundFile==0.10.3.post1
tensorboard==2.2.1
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
threadpoolctl==2.0.0
tifffile==2020.5.11
tqdm==4.46.0
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
